Add destination parameter to convert_to_parquet function
The convert_to_parquet function now takes an additional parameter for the destination directory where parquet files are to be stored. This enables the user to specify the directory for their output files instead of overwriting the input files. The changes have been reflected in the parameter list, the function body, and all function calls.